Added invalidation on Edit and Delete.

+++ Services/Administration/OJS.Services.Administration.Business/Tests/TestsBusinessService.cs
@@ -30,6 +30,7 @@ public class TestsBusinessService : AdministrationOperationService<Test, int, Te
     private readonly ISubmissionsDataService submissionsDataService;
     private readonly IProblemsBusinessService problemsBusinessService;
     private readonly IZipArchivesService zipArchivesService;
+    private readonly ITestsCacheService testsCache;
 
     public TestsBusinessService(
         ITestsDataService testsDataService,
@@ -39,7 +40,8 @@ public class TestsBusinessService : AdministrationOperationService<Test, int, Te
         IZippedTestsParserService zippedTestsParserService,
         ISubmissionsDataService submissionsDataService,
         IProblemsBusinessService problemsBusinessService,
-        IZipArchivesService zipArchivesService)
+        IZipArchivesService zipArchivesService,
+        ITestsCacheService testsCache)
     {
         this.testsDataService = testsDataService;
         this.testRunsDataService = testRunsDataService;
@@ -49,6 +51,7 @@ public class TestsBusinessService : AdministrationOperationService<Test, int, Te
         this.submissionsDataService = submissionsDataService;
         this.problemsBusinessService = problemsBusinessService;
         this.zipArchivesService = zipArchivesService;
+        this.testsCache = testsCache;
     }
 
     public override async Task<TestAdministrationModel> Get(int id)
@@ -91,6 +94,8 @@ public class TestsBusinessService : AdministrationOperationService<Test, int, Te
 
         test.MapFrom(model);
 
+        await this.testsCache.ClearTestsCacheByProblemId(test.ProblemId);
+
         UpdateInputAndOutput(test, model);
         UpdateType(test, model);
 
@@ -116,6 +121,17 @@ public class TestsBusinessService : AdministrationOperationService<Test, int, Te
 
     public override async Task Delete(int id)
     {
+        var test = await this.testsDataService
+            .GetByIdQuery(id)
+            .Select(t => new { t.ProblemId })
+            .FirstOrDefaultAsync();
+
+        if (test is null)
+        {
+            return;
+        }
+
+        await this.testsCache.ClearTestsCacheByProblemId(test.ProblemId);
         await this.testsDataService.DeleteById(id);
         await this.testsDataService.SaveChanges();
     }
@@ -123,6 +139,7 @@ public class TestsBusinessService : AdministrationOperationService<Test, int, Te
     public async Task DeleteAll(int problemId)
     {
         await this.testRunsDataService.DeleteByProblem(problemId);
+        await this.testsCache.ClearTestsCacheByProblemId(problemId);
         await this.testsDataService.DeleteByProblem(problemId);
     }
